15th 1.more things added to user-detail.jsp wch says wat links blogs a perticulr user has. we get lazy initialization exception bcause in usecontrlr we slect data from table user,and user.java has blogs and each blog contains items and this mapping is by default lazy, @OneTOMany(mappedBy="user",fetch=FetchType.LAZY) private List<Blog> blogs;
what does it mean by lazy?it means whn we load this enity user frm db
we wont load this blog with this user instead there ll be some proxy for
blog,and when we load this outside the transaction then we ll get this
lazyinitializationexception ,so in user-detail.jsp we r accesssing
${user.blogs} outside the transaction which is in the presentation
layer.

how to overcome this?
there are two ways,
the first easy way is to use af ilter wch wll open a traction in
presentation layr,bt it hs compliction that we dnt have contrl over db
compltly ,
1 way .add this filter in web.xml

<filter>
    <filter-name>oemInViewFilter</filter-name>
    <filter-class>
        org.springframework.orm.jpa.support.OpenEntityManagerInViewFilter
    </filter-class>
    <init-param>
    <param-name>entityManagerFactoryBeanName</param-name>
    <param-value>emf</param-value>
    </init-param>
</filter>

<filter-mapping>
    <filter-name>oemInViewFilter</filter-name>
    <url-pattern>/*</url-pattern>
</filter-mapping>